refactor: DTO 위치 변경
