+++ saerojinro-admin/src/main/java/goorm/saerojinro/admin/notification/application/NotificationAdminFacade.java
@@ -6,9 +6,8 @@ import goorm.saerojinro.domain.notification.domain.Notification;
 import goorm.saerojinro.domain.reservation.application.ReservationQueryService;
 import goorm.saerojinro.domain.reservation.domain.Reservation;
 import goorm.saerojinro.domain.user.application.UserQueryService;
-import goorm.saerojinro.domain.user.domain.User;
-import goorm.saerojinro.infra.notification.exception.EmitterNotFoundException;
-import goorm.saerojinro.infra.notification.request.NotificationSendRequest;
+import goorm.saerojinro.admin.notification.exception.EmitterNotFoundException;
+import goorm.saerojinro.admin.notification.presentation.request.NotificationSendRequest;
 import goorm.saerojinro.infra.notification.response.NotificationSendResponse;
 import goorm.saerojinro.infra.notification.sse.NotificationSseSender;
 import lombok.RequiredArgsConstructor;
